requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
msal==1.31.0
python-dotenv==1.0.1
itsdangerous==2.2.0
aiohttp==3.10.5
cryptography==42.0.5
asyncpg
pypdf
docx2txt
python-pptx
langchain-openai
langchain-text-splitters
